+++ CHANGES.md
@@ -1,3 +1,8 @@
+## 0.3.0
+- added `--pipe`
+- `--watch` can now accept paths as arguments. crescent will `--compile` and `--run` any given paths when a change in these paths are found. `--watch` without arguments defaults to the old behavior of watching every path given with `--compile` or `--run`, while only processing relevant files.
+- went back to the old single-string help format - any higher level of complexity is unneeded.
+
 ## 0.2.0
 - `--run` is now a cli option, as opposed to crescent working off of arguments
 - added `--no-colors`